matches_app/requirements.txt
fastapi
tensorflow
keras
sqlalchemy
pandas
typing
uvicorn
python-dotenv==1.0.0
psycopg2
psycopg2-binary
PyJWT